+++ src/pages/Game/Game.helpers.ts
@@ -1,0 +1,53 @@
+import React, {useState} from "react";
+import {useHistory} from "react-router-dom";
+
+import {World} from "@/game/world/world";
+import {routes} from "@/config/routes/routes";
+
+export const useGame = () => {
+    const [active, setActive] = useState(true);
+    const [pause, setPause] = useState(false);
+    const [world] = useState(new World());
+    const canvasRef = React.useRef<HTMLCanvasElement>(null);
+    const history = useHistory();
+
+    const callMenu = (e: KeyboardEvent) => {
+        if (e.key === 'Escape') {
+            setActive(true);
+            setPause(true);
+            world.setPause(true);
+        }
+    }
+
+    const setUpPauseButton = () => {
+        document.addEventListener('keydown', callMenu);
+    }
+
+    const onClose = () => {
+        setActive(false);
+        world.destroy()
+        history.push(`/#${routes.main.path}`);
+    };
+
+    const onStart = () => {
+        world.init(canvasRef.current);
+        setPause(false);
+        setActive(false);
+    }
+
+    const onResume = () => {
+        world.setPause(false);
+        setPause(false);
+        setActive(false);
+    }
+
+    return {
+        active,
+        pause,
+        canvasRef,
+        onStart,
+        onResume,
+        onClose,
+        setUpPauseButton
+    }
+}
