+++ src/pages/Game/Menu/Main.module.scss
@@ -1,9 +1,11 @@
+@use './src/styles/variables.module.scss' as variables;
+
 .buttonContainer {
   display: flex;
   flex-direction: column;
-  padding: 24px;
+  padding: variables.$spacing_x2;
 
   > button {
-    margin-bottom: 24px;
+    margin-bottom: variables.$spacing_x2;
   }
 }
